+++ spec/utils/postal_spec.js
@@ -2,14 +2,21 @@ import h from 'spec/spec_helper';
 import { lazy_require } from 'azk';
 
 var l = lazy_require({
-  IPublisher: ['azk/utils/postal'],
-  subscribe : ['azk/utils/postal'],
-  publish   : ['azk/utils/postal'],
+  IPublisher    : ['azk/utils/postal'],
+  channel       : ['azk/utils/postal'],
+  subscribe     : ['azk/utils/postal'],
+  publish       : ['azk/utils/postal'],
+  subscriptions : ['azk/utils/postal'],
+  unsubscribeAll: ['azk/utils/postal'],
 });
 
 describe("Azk utils, postal module", function() {
+  beforeEach(() => {
+    l.channel.unsubscribeAll();
+  });
+
   it("should support subscribe and publish with defailt channel", function(done) {
-    var topic = 'test.postal.module';
+    var topic = "test.postal.module";
     var data  = { data: true };
     var subs  = l.subscribe(topic, (msg) => {
       h.expect(msg).to.eql(data);
@@ -18,6 +25,23 @@ describe("Azk utils, postal module", function() {
     });
 
     l.publish(topic, data);
+  });
+
+  it("should return all subscribe for default channel if call subscriptions", function() {
+    var topic = "test.subscriptions";
+    var sub   = l.subscribe(topic, () => {});
+    var subs  = l.subscriptions();
+    h.expect(subs).to.eql([sub]);
+  });
+
+  it("should support unsubscribe all subscriptions", function() {
+    var topic = "test.subscriptions";
+    var sub   = l.subscribe(topic, () => {});
+    var subs  = l.subscriptions();
+    h.expect(subs).to.eql([sub]);
+    l.unsubscribeAll();
+    subs = l.subscriptions();
+    h.expect(subs).to.not.eql([sub]);
   });
 
   it("should raise erro if subscribe without topic", function() {
@@ -50,6 +74,19 @@ describe("Azk utils, postal module", function() {
       });
 
       var obj = new MyClass();
+      obj.test_publish('subtopic', data);
+    });
+
+    it("should support subscribe with default prefix topic", function(done) {
+      var data = { data: true };
+      var obj  = new MyClass();
+
+      var subs = obj.subscribe((msg) => {
+        h.expect(msg).to.eql(data);
+        subs.unsubscribe();
+        done();
+      });
+
       obj.test_publish('subtopic', data);
     });
   });