+++ src/system/scale.js
@@ -1,4 +1,4 @@
-import { Q, async, _, lazy_require } from 'azk';
+import { Q, async, _, lazy_require, publish } from 'azk';
 import { calculateHash } from 'azk/utils';
 import { SystemDependError, SystemNotScalable } from 'azk/utils/errors';
 import { Balancer } from 'azk/system/balancer';
@@ -27,7 +27,7 @@ var Scale = {
       dependencies: true,
     });
 
-    return async(this, function* (notify) {
+    return async(this, function* () {
       // how many times the
       var containers = yield this.instances(system);
 
@@ -42,7 +42,7 @@ var Scale = {
       }
 
       if (icc !== 0) {
-        notify({ type: "scale", from: from, to: from + icc, system: system.name });
+        publish("system.scale.status", { type: "scale", from: from, to: from + icc, system: system.name });
       }
 
       if (icc > 0) {